refactor: add missing imports
