+++ src/renderer/components/Progressbar/index.tsx
@@ -2,11 +2,17 @@ import * as React from 'react';
 
 import { StyledProgressbar, Line, Track } from './style';
 
-export const Progressbar = ({ value, style }: { value: number, style?: React.CSSProperties }) => {
+export const Progressbar = ({
+  value,
+  style,
+}: {
+  value: number;
+  style?: React.CSSProperties;
+}) => {
   return (
     <StyledProgressbar style={style}>
       <Line />
       <Track style={{ width: `${value}%` }} />
     </StyledProgressbar>
   );
-}
+};
